refactor(viewer): extract export and zoom managers
Extract ExportManager and ZoomManager from SquiggleViewer to improve
code organization and maintainability. The viewer.py file is reduced
from 2,291 to 1,809 lines (21% reduction).

Managers extracted:
- ExportManager (349 lines): Handles HTML, PNG, SVG plot export with
  dimension control and current zoom level support
- ZoomManager (211 lines): Manages JavaScript-based zoom/pan operations,
  range saving/restoration, and sequence match zooming

The viewer delegates to managers via thin wrapper methods, maintaining
backward compatibility. All 418 tests pass.
